+++ src/pages/Home.jsx
@@ -1,9 +1,8 @@
-import { useNavigate } from "react-router-dom";
 import { POST_CONTENT, POST_TITLE } from "@data/HomePostCardContent";
 import PostCard from "@components/home/PostCard";
+import TestButton from "../components/common/TestButton";
 
 const Home = () => {
-  const navigate = useNavigate();
   return (
     <div className="col-base justify-center my-20 w-full">
       <div className="pb-20">
@@ -20,16 +19,7 @@ const Home = () => {
           {POST_CONTENT.EFFECT_CONTENT_2}
         </PostCard>
       </div>
-      <div className="flex my-20 justify-center">
-        <button
-          onClick={() => {
-            navigate("/test");
-          }}
-          className="btn-base h-sm w-lg shadow-lg border-2 "
-        >
-          테스트하러 가기
-        </button>
-      </div>
+      <TestButton />
     </div>
   );
 };
